Revert move of legacy functions from environments into separate module.
This commit reverts the previous refactorization where a lot of the
legacy methods of environments were moved into the legacy_templating
module. This refactorization turned out to lead to issues in combination
with environments derived from these classes.
