+++ flow/legacy_templating.py
@@ -6,7 +6,6 @@
 Will be removed beginning version 0.8.
 """
 from __future__ import print_function
-import sys
 import io
 import warnings
 import functools
@@ -302,145 +301,3 @@ def support_legacy_templating_submit(func):
         return func(self, script, *args, **kwargs)
 
     return wrapper
-
-
-def support_legacy_templating(cls):
-
-    if cls.__name__ == 'TestEnvironment':
-
-        @classmethod
-        @deprecated_since_06
-        def mpi_cmd(cls, cmd, np):
-            return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
-        cls.mpi_cmd = mpi_cmd
-
-        @classmethod
-        @deprecated_since_06
-        def script(cls, **kwargs):
-            js = super(cls, cls).script(**kwargs)
-            for key in sorted(kwargs):
-                js.writeline('#TEST {}={}'.format(key, kwargs[key]))
-            return js
-        cls.script = script
-
-    elif cls.__name__ == 'NodesEnvironment':
-        from .errors import SubmitError
-
-        @classmethod
-        @deprecated_since_06
-        def calc_num_nodes(cls, np_total, ppn, force=False, **kwargs):
-            from math import ceil
-            if ppn is None:
-                try:
-                    ppn = getattr(cls, 'cores_per_node')
-                except AttributeError:
-                    raise SubmitError(NUM_NODES_WARNING)
-
-            # Calculate the total number of required nodes
-            nn = int(ceil(np_total / ppn))
-
-            if not force:  # Perform basic check concerning the node utilization.
-                usage = np_total / nn / ppn
-                if usage < 0.9:
-                    print(UTILIZATION_WARNING.format(ppn=ppn, usage=usage), file=sys.stderr)
-                    raise SubmitError("Bad node utilization!")
-            return nn
-        cls.calc_num_nodes = calc_num_nodes
-
-        @classmethod
-        @deprecated_since_06
-        def add_args(cls, parser):
-            super(cls, cls).add_args(parser)
-            parser.add_argument(
-                '--nn',
-                type=int,
-                help="Specify the total number of nodes. This value is computed automatically "
-                     "from the 'np' directive otherwise.")
-        cls.add_args = add_args
-
-    elif cls.__name__ == 'DefaultTorqueEnvironment':
-
-        @classmethod
-        @deprecated_since_06
-        def mpi_cmd(cls, cmd, np):
-            return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
-        cls.mpi_cmd = mpi_cmd
-
-        @classmethod
-        @deprecated_since_06
-        def gen_tasks(cls, js, np_total):
-            """Helper function to generate the number of tasks (for overriding)"""
-            js.writeline('#PBS -l nodes={}'.format(np_total))
-            return js
-        cls.gen_tasks = gen_tasks
-
-        @classmethod
-        @deprecated_since_06
-        def script(cls, _id, np_total, walltime=None, no_copy_env=False, **kwargs):
-            from .util.template_filters import format_timedelta
-            js = super(cls, cls).script()
-            js.writeline('#PBS -N {}'.format(_id))
-            js = cls.gen_tasks(js)
-            if walltime is not None:
-                js.writeline('#PBS -l walltime={}'.format(format_timedelta(walltime)))
-            if not no_copy_env:
-                js.writeline('#PBS -V')
-            return js
-        cls.script = script
-
-    elif cls.__name__ == 'DefaultSlurmEnvironment':
-
-        @classmethod
-        @deprecated_since_06
-        def mpi_cmd(cls, cmd, np):
-            return 'mpirun -np {np} {cmd}'.format(np=np, cmd=cmd)
-        cls.mpi_cmd = mpi_cmd
-
-        @classmethod
-        @deprecated_since_06
-        def gen_tasks(cls, js, np_total):
-            """Helper function to generate the number of tasks (for overriding)"""
-            js.writeline('#SBATCH --ntasks={}'.format(np_total))
-            return js
-        cls.gen_tasks = gen_tasks
-
-        @classmethod
-        @deprecated_since_06
-        def script(cls, _id, np_total, walltime=None, **kwargs):
-            from .util.template_filters import format_timedelta
-            js = super(cls, cls).script()
-            js.writeline('#!/bin/bash')
-            js.writeline('#SBATCH --job-name="{}"'.format(_id))
-            js = cls.gen_tasks(js, np_total)
-            if walltime is not None:
-                js.writeline('#SBATCH -t {}'.format(format_timedelta(walltime)))
-            return js
-        cls.script = script
-
-    elif cls.__name__ == 'DefaultLSFEnvironment':
-
-        @classmethod
-        @deprecated_since_06
-        def mpi_cmd(cls, cmd, np):
-            raise NotImplementedError("LSF environments are not supported by the "
-                                      "legacy templating system.")
-        cls.mpi_cmd = mpi_cmd
-
-        @classmethod
-        @deprecated_since_06
-        def gen_tasks(cls, js, np_total):
-            raise NotImplementedError("LSF environments are not supported by the "
-                                      "legacy templating system.")
-        cls.gen_tasks = gen_tasks
-
-        @classmethod
-        @deprecated_since_06
-        def script(cls, _id, np_total, walltime=None, **kwargs):
-            raise NotImplementedError("LSF environments are not supported by the "
-                                      "legacy templating system.")
-        cls.script = script
-
-    else:
-        raise TypeError("No legacy templating defined for '{}'.".format(cls))
-
-    return cls
